Add vehicle dropdown select

+++ frontend/src/pages/HomePage.tsx
@@ -26,6 +26,9 @@ const defaultSettings: Settings = {
     darkTheme: false,
     units: "miles"
 };
+interface Vehicles {
+    [make: string]: string[];
+}
 
 function HomePage() {
     const [menuOpen, setMenuOpen] = useState(false);
@@ -57,6 +60,7 @@ function HomePage() {
     const [password, setPassword] = useState("");
     const [userID, setUserID] = useState("");
     const [tripID, setTripID] = useState(-1);
+    const [vehicleData, setVehicleData] = useState<Vehicles | null>(null);
     const [vehicleMake, setVehicleMake] = useState(localStorage.getItem("vehicleMake") || "");
     const [vehicleModel, setVehicleModel] = useState(localStorage.getItem("vehicleModel") || "");
 
@@ -139,6 +143,11 @@ function HomePage() {
       return () => window.removeEventListener("click", startAudio);
     }, []);
 
+    // perform once when webpage loads
+    useEffect(() => {
+        loadVehicleData();
+    }, []);
+
  
     useEffect(() => {
       if (savedOpen && userID) {
@@ -162,6 +171,27 @@ function HomePage() {
         getTrips();
       }
     }, [savedOpen, userID]);
+
+    const loadVehicleData = async () => {
+        if (vehicleData || isLoading) {
+            return;
+        } 
+
+        setIsLoading(true);
+        const data = await import("./vehicles.json");
+        setVehicleData(data.default);
+        setIsLoading(false);
+    }
+
+    const vehicleMakeOptions = vehicleData ? [{ value: "", label: ""}, ...Object.keys(vehicleData).map((make) => ({
+        value: make,
+        label: make
+    }))] : [];
+
+    const vehicleModelOptions = vehicleMake && vehicleData ? [{ value: "", label: ""}, ...vehicleData[vehicleMake].map((model) => ({
+        value: model,
+        label: model
+    }))] : [];
 
 
     const handleGetUser = async () => {
@@ -360,17 +390,21 @@ function HomePage() {
     const handleSaveTrip = async (name: string, isNew: boolean) => {
         // if (!name.trim()) return;
         // No trip selected (tripID starts at 1)
-        if (tripID <= 0 || isNew) {
-            console.log("here");
-            const newTripID: number = await handleAddTrip(name);
-            await handleAddStops(newTripID);
+        if (userID) {
+            if (tripID <= 0 || isNew) {
+                console.log("here");
+                const newTripID: number = await handleAddTrip(name);
+                await handleAddStops(newTripID);
+            } else {
+                console.log("saving");
+                await handleDeleteStops(tripID);
+                await handleAddStops(tripID);
+            }
+            console.log("Trip saved!");
         } else {
-            console.log("saving")
-            await handleDeleteStops(tripID);
-            await handleAddStops(tripID);
+            alert("You must be logged in to save trips.");
         }
         // save trip logic goes here
-        console.log("Trip saved!");
         setSaveTripOpen(false);
     };
 
@@ -802,22 +836,29 @@ function HomePage() {
             </WindowHeader>
             <WindowContent>
                 <p>Enter your vehicle's make (brand):</p>
-                <TextInput
-                placeholder="Make Name"
-                value={vehicleMake}
-                onChange={(e) => setVehicleMake(e.target.value)}
-                fullWidth
-                autoFocus
-                style={{ marginTop: 10, marginBottom: 20 }}
+                <Select
+                    onFocus={loadVehicleData}
+                    defaultValue={vehicleMake}
+                    value={vehicleMake}
+                    options={vehicleMakeOptions}
+                    menuMaxHeight={160}
+                    width={160}
+                    onChange={(e) => {
+                        setVehicleMake(e.value);
+                        setVehicleModel(""); // Reset model
+                    }}
                 />
                 <p>Enter your vehicle's model:</p>
-                <TextInput
-                placeholder="Model Name"
-                value={vehicleModel}
-                onChange={(e) => setVehicleModel(e.target.value)}
-                fullWidth
-                autoFocus
-                style={{ marginTop: 10, marginBottom: 20 }}
+                <Select
+                    defaultValue={vehicleModel}
+                    disabled={!vehicleMake}
+                    value={vehicleModel}
+                    options={vehicleModelOptions}
+                    menuMaxHeight={160}
+                    width={160}
+                    onChange={(e) => {
+                        setVehicleModel(e.value);
+                    }}
                 />
                 {/* PUT VEHICLE API CONFIRM HERE AND CHANGE CALL TO USE VARS*/}
                 <Button
@@ -1085,7 +1126,7 @@ function HomePage() {
                                 <div style={{ marginTop: "25%", display: "flex", justifyContent: "space-between", width: "100%" }}>
                                     <Button
                                         disabled={isLoading}
-                                        onClick={handleGetDirections()}
+                                        onClick={() => handleGetDirections()}
                                     >
                                         Confirm
                                     </Button>
